Appended the values into the named tuple.
By printing the match.group(n) i was able to see the type of values that would be printed and for each of the groups was able to apply any changes that needed to be done before adding it to the named tuple.

date i wanted to ensure that the month and year were both present and also ensure that if the month was dec then the next month would be jan and the year would change eg dec 2023 -> jan 2024 so needed to record that value so that when it comes to analysing the data it would be easier.

transaction type and transaction details don't need any changes made to them as the full string of them would be needed.

withdrawal and deposit is a bit of an issue still as I haven't found a way to separate it so that there is a separate value in the tuple for deposit and a separate value for withdrawal

balance has been added as well.

before converting to csv will need to separate Withdrawal and Deposit. possible idea could be to  see if the balance of the previous value was less than or greater than and then could find out if deposit or withdrawal.