Updated the Withdrawals and Deposits
Updated Withdrawals and Deposits columns so that it takes float data type instead of string,

Then i dropeed the WD column as now I have separated the withdrawal and deposit columns and updated the dataframe to include those two columns.

Also filled all null values with 0 which is another reason for updating the Deposit and Withdrawal to be floating values.