Conversion of all data types and creation of CSV
finished the conversion of all the columns and then returned the dataframe. Can now create a CSV out of it or it is possible to just create some graphs using pandas. Maybe try integrating a small UI to allow the pdf to be uploaded as opposed to the PDF file being submitted via file path.