Created the CSV file
Created the CSV file, opened it in excel and deleted the first row as it just contained the counter which is already present in excel.

Created a pivot table to analyse my monthly spending amounts. Realised that spending is quite high as all my spending would be equivilant to what is leaving the bank account and after looking through my statement again i realised that spending shouldn't be that high as a lot of the time when i get paid im transferring that income to my other savings account however as it is considered a withdrawal (leaving the account) it is counting it as spending.

will need to fix it so that rows in which i am transferring the money are not included in the dataset.