Fixed it so that when im transferring my money into savings
Fixed the issue so that it doesnt pick up when i transfer money from my spending to saving account via the creation of a new regex statement called remove_pattern which detects the transfer via picking up the account number which is then followed by "DEBIT TRANSFER" eg "00-0000-0000000-00 DEBIT TRANSFER".

And then when I append to the bank_line_items i included a conditional statement so that if the remove pattern is detected then it won't append that row into the bank_line_items list.